LPS-58809 move to module
