+++ modules/apps/export-import/export-import-service/src/com/liferay/exportimport/background/task/PortletExportImportBackgroundTaskStatusMessageTranslator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.exportimport.backgroundtask;
+package com.liferay.exportimport.background.task;
 
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskStatus;
 import com.liferay.portal.kernel.messaging.Message;
