+++ modules/apps/export-import/export-import-service/src/com/liferay/exportimport/background/task/PortletImportBackgroundTaskExecutor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.exportimport.backgroundtask;
+package com.liferay.exportimport.background.task;
 
 import com.liferay.portal.kernel.backgroundtask.BackgroundTask;
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskConstants;
@@ -37,12 +37,12 @@ import java.util.concurrent.Callable;
  * @author Daniel Kocsis
  * @author Akos Thurzo
  */
-public class LayoutImportBackgroundTaskExecutor
+public class PortletImportBackgroundTaskExecutor
 	extends BaseExportImportBackgroundTaskExecutor {
 
-	public LayoutImportBackgroundTaskExecutor() {
+	public PortletImportBackgroundTaskExecutor() {
 		setBackgroundTaskStatusMessageTranslator(
-			new LayoutExportImportBackgroundTaskStatusMessageTranslator());
+			new PortletExportImportBackgroundTaskStatusMessageTranslator());
 
 		// Isolation level guarantees this will be serial in a group
 
@@ -51,17 +51,18 @@ public class LayoutImportBackgroundTaskExecutor
 
 	@Override
 	public BackgroundTaskExecutor clone() {
-		LayoutImportBackgroundTaskExecutor layoutImportBackgroundTaskExecutor =
-			new LayoutImportBackgroundTaskExecutor();
+		PortletImportBackgroundTaskExecutor
+			portletImportBackgroundTaskExecutor =
+				new PortletImportBackgroundTaskExecutor();
 
-		layoutImportBackgroundTaskExecutor.
+		portletImportBackgroundTaskExecutor.
 			setBackgroundTaskStatusMessageTranslator(
 				getBackgroundTaskStatusMessageTranslator());
 
-		layoutImportBackgroundTaskExecutor.setIsolationLevel(
+		portletImportBackgroundTaskExecutor.setIsolationLevel(
 			getIsolationLevel());
 
-		return layoutImportBackgroundTaskExecutor;
+		return portletImportBackgroundTaskExecutor;
 	}
 
 	@Override
@@ -84,14 +85,14 @@ public class LayoutImportBackgroundTaskExecutor
 
 				TransactionInvokerUtil.invoke(
 					transactionAttribute,
-					new LayoutImportCallable(exportImportConfiguration, file));
+					new PortletImportCallable(exportImportConfiguration, file));
 			}
 			catch (Throwable t) {
 				if (_log.isDebugEnabled()) {
 					_log.debug(t, t);
 				}
 				else if (_log.isWarnEnabled()) {
-					_log.warn("Unable to import layouts: " + t.getMessage());
+					_log.warn("Unable to import portlet: " + t.getMessage());
 				}
 
 				throw new SystemException(t);
@@ -105,11 +106,11 @@ public class LayoutImportBackgroundTaskExecutor
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		LayoutImportBackgroundTaskExecutor.class);
+		PortletImportBackgroundTaskExecutor.class);
 
-	private class LayoutImportCallable implements Callable<Void> {
+	private class PortletImportCallable implements Callable<Void> {
 
-		public LayoutImportCallable(
+		public PortletImportCallable(
 			ExportImportConfiguration exportImportConfiguration, File file) {
 
 			_exportImportConfiguration = exportImportConfiguration;
@@ -118,10 +119,10 @@ public class LayoutImportBackgroundTaskExecutor
 
 		@Override
 		public Void call() throws PortalException {
-			ExportImportLocalServiceUtil.importLayoutsDataDeletions(
+			ExportImportLocalServiceUtil.importPortletDataDeletions(
 				_exportImportConfiguration, _file);
 
-			ExportImportLocalServiceUtil.importLayouts(
+			ExportImportLocalServiceUtil.importPortletInfo(
 				_exportImportConfiguration, _file);
 
 			return null;
